Moved serial_connection directory into device directory
