+++ src-tauri/src/constructors/mock/mocks.rs
@@ -5,8 +5,8 @@ use crate::constructors::init::init_edge_map::as_key;
 use crate::data_conversion::distance_conversion::{
     get_spherical_distance, gps_degrees_to_protobuf_field,
 };
-use crate::mesh::device::helpers::get_current_time_u32;
-use crate::mesh::device::MeshNode;
+use crate::device::helpers::get_current_time_u32;
+use crate::device::MeshNode;
 use app::protobufs;
 use rand::distributions::Uniform;
 use rand::prelude::*;
