+++ src-tauri/src/device/helpers.rs
