+++ src-tauri/src/device/serial_connection/helpers.rs
