+++ src-tauri/src/device/serial_connection/mod.rs
