--- src-tauri/src/mesh/mod.rs
@@ -1,2 +1,0 @@
-pub mod device;
-pub mod serial_connection;
